Limit cpu usage of the service inside engine appliance
As the service runs an ifinite loop, it can consume all the
available cpu resource. It won't matter much if the proccessing
speed of this service is slower. So the solution can be using
linux's cpu affinity to limit jvm thread to a subset of cpus.

This patch bind service's thread to 1 cpu thread out of every 4 cpu
threads.

Signed-off-by: walteryang <walteryang47@gmail.com>